Numpyifies code: Data structure changed to enhance performance
