+++ cell.py
@@ -1,22 +1,18 @@
 #!/usr/bin/python  -u
 
-PYTHONUNBUFFERED = 1
-
-import argparse
-import collections
-import copy
-import gzip
 import itertools
-import os
-import pickle
 import sys
+import warnings
 from math import exp, log, sqrt
 
 import numpy as np
-from mayavi import mlab
-from numpy.random import random
-
+import numpy.random as npr
+import numpy.ma as ma
+import scipy.linalg
 import voronoi_neighbors
+from mayavi import mlab
+
+warnings.filterwarnings("ignore", category=DeprecationWarning)
 
 null = np.array([0.0, 0.0, 0.0])
 ex = np.array([1.0, 0.0, 0.0])
@@ -24,39 +20,54 @@ ey = np.array([0.0, 1.0, 0.0])
 ez = np.array([0.0, 0.0, 1.0])
 
 
-def norm(v, what=None):
-    n = np.sqrt(np.dot(v, v))
-    if n > 1e-6:
-        u = v / n
-    else:
-        u, n = null, 0.
-    if what == 'vec':
-        return u
-    if what == 'mag' or what == 'len':
-        return n
-    return [u, n]
-
-
 def ccw(A, B, C):
     return (C[1] - A[1]) * (B[0] - A[0]) > (B[1] - A[1]) * (C[0] - A[0])
 
 
+def getNormvec(v):
+    # returns normalized v
+    d = scipy.linalg.norm(v, axis=-1)
+    v_masked = ma.getdata(ma.array(v) / ma.array(d[..., None]))
+    v = ma.getdata(v_masked.filled(0))
+
+    return v
+
+
+def getNormtoo(v):
+    # returns norm of v and normalized v
+    d = scipy.linalg.norm(v, axis=-1)
+    v_masked = ma.array(v) / ma.array(d[..., None])
+    v = ma.getdata(v_masked.filled(0))
+
+    return v, d
+
+
 def intersect(A, B, C, D):
-    if norm(A - C, "mag") < 0.01: return False
-    if norm(A - D, "mag") < 0.01: return False
-    if norm(B - C, "mag") < 0.01: return False
+    mynorm = scipy.linalg.norm(np.array([A - C, A - D, B - C]), axis=1)
+    if (mynorm < 0.01).any():
+        return False
     return ccw(A, C, D) != ccw(B, C, D) and ccw(A, B, C) != ccw(A, B, D)
 
 
-def R(dphi, v):
-    axis, theta = norm(dphi)
-    if theta < 1e-6:
-        return v
-    a = np.cos(theta / 2)
-    b, c, d = axis * np.sin(theta / 2)
-    return np.dot(np.array([[a * a + b * b - c * c - d * d, 2 * (b * c - a * d), 2 * (b * d + a * c)],
-                            [2 * (b * c + a * d), a * a + c * c - b * b - d * d, 2 * (c * d - a * b)],
-                            [2 * (b * d - a * c), 2 * (c * d + a * b), a * a + d * d - b * b - c * c]]), v)
+def getRotMatArray(Phis):
+    Thetas = scipy.linalg.norm(Phis, axis=1)
+    Axes_masked = ma.array(Phis) / ma.array(Thetas[..., None])
+    Axes = ma.getdata(Axes_masked.filled(0))
+    a = np.cos(Thetas / 2)
+    b, c, d = np.transpose(Axes) * np.sin(Thetas / 2)
+    RotMat = np.array([[a * a + b * b - c * c - d * d, 2 * (b * c - a * d), 2 * (b * d + a * c)],
+                      [2 * (b * c + a * d), a * a + c * c - b * b - d * d, 2 * (c * d - a * b)],
+                      [2 * (b * d - a * c), 2 * (c * d + a * b), a * a + d * d - b * b - c * c]])
+    return np.transpose(RotMat, axes=(2, 0, 1))
+
+
+def getRotMat(Phis):
+    Axis, Theta = getNormtoo(Phis)
+    a = np.cos(Theta / 2)
+    b, c, d = Axis * np.sin(Theta / 2)
+    return np.array([[a * a + b * b - c * c - d * d, 2 * (b * c - a * d), 2 * (b * d + a * c)],
+                    [2 * (b * c + a * d), a * a + c * c - b * b - d * d, 2 * (c * d - a * b)],
+                    [2 * (b * d - a * c), 2 * (c * d + a * b), a * a + d * d - b * b - c * c]])
 
 
 def update_progress(progress):
@@ -80,25 +91,75 @@ def update_progress(progress):
         status = "Done...\r\n"
     block = int(round(barLength * progress))
     text = "\rProgress: [{0}] {1} % {2}".format("#" * block + "-" * (barLength - block), round(progress * 100, 1),
-                                               status)
+                                                status)
     sys.stdout.write(text)
     sys.stdout.flush()
 
 
-class Configuration():
-    def __init__(self, num=None, dumpdir=None, dt=0.01, nmax=3000, qmin=0.001, d0_0=1, force_limit=15., p_add=1.,
-                 p_del=0.2, chkx=True, anis=0.0, d0max=2., d2=False, unrestricted=True, distlimit=3):
-        self.nodes = []
-        self.links = []
-        self.nNodes = 0  # bookkeeping for state vector position (I don't get that)
-        self.dumpdir = dumpdir  # string for directory to save configs
+def showconfig(configs, links, nodeforces, fl, figure=None, figureindex = 0, bgcolor=(1, 1, 1), fgcolor=(0, 0, 0),
+               figsize=(1000, 1000), cmap='viridis', vmaxlinks=5, vmaxcells=5, cbar=False):
+    if figure is None:
+         fig = mlab.figure(figureindex, bgcolor=bgcolor, fgcolor=fgcolor, size=figsize)
+    else:
+         fig = figure
+    x, y, z = configs.T
+    xl, yl, zl = configs[links[..., 0]].T
+    rxl, ryl, rzl = (configs[links[..., 1]] - configs[links[..., 0]]).T
+    fc = scipy.linalg.norm(nodeforces, axis=1)
+
+    cells = mlab.points3d(x, y, z, fc, scale_factor=1, opacity=0.5, resolution=16, scale_mode='none', vmin=0.,
+                          colormap=cmap, vmax=vmaxcells)
+    links = mlab.quiver3d(xl, yl, zl, rxl, ryl, rzl, scalars=fl, mode='2ddash', line_width=4., scale_mode='vector',
+                          scale_factor=1, colormap=cmap, vmin=0., vmax=vmaxlinks)
+    links.glyph.color_mode = "color_by_scalar"
+    if cbar:
+        mlab.scalarbar(links, nb_labels=2, title='Force on link')
+    return cells, links
+
+
+@mlab.animate(delay=500)
+def animateconfigs(Configs, Links, nodeForces, linkForces, ts, figureindex=0, bgcolor=(1, 1, 1),
+                   fgcolor=(0, 0, 0), figsize=(1000, 1000), cmap='viridis', cbar=False):
+    fig = mlab.figure(figureindex, bgcolor=bgcolor, fgcolor=fgcolor, size=figsize)
+    vmaxcells = np.max(scipy.linalg.norm(nodeForces, axis=2))
+    vmaxlinks = max([np.max(timestep) for timestep in linkForces])
+    cells, links = showconfig(Configs[0], Links[0], nodeForces[0], linkForces[0], fig,
+                              vmaxcells=vmaxcells, vmaxlinks=vmaxlinks)
+    text = mlab.title('0.0', height=.9)
+
+    while True:
+        for (c, l, nF, fl, t) in zip(Configs, Links, nodeForces, linkForces, ts):
+            x, y, z = c.T
+            xl, yl, zl = c[l[..., 0]].T
+            rxl, ryl, rzl = (c[l[..., 1]] - c[l[..., 0]]).T
+            fc = scipy.linalg.norm(nF, axis=1)
+
+            cells.mlab_source.set(x=x, y=y, z=z, scalars=fc)
+            links.mlab_source.reset(x=xl, y=yl, z=zl, u=rxl, v=ryl, w=rzl, scalars=fl)
+            text.set(text='{}'.format(round(t, 2)))
+            # print 'Updating... '
+            yield
+
+class Configuration:
+    def __init__(self, num, dt=0.01, nmax=3000, qmin=0.001, d0_0=1, force_limit=15., p_add=1.,
+                 p_del=0.2, chkx=False, anis=0.0, d0max=2., dims=3):
+        if dims == 2:
+            self.updateLinkForces = lambda PHI, T, Norm, NormT, Bend, Twist, K, D0, Nodeinds: \
+                self.updateLinkForces2D(PHI, T, Norm, NormT, Bend, Twist, K, D0, Nodeinds)
+            self.dims = dims
+        elif dims == 3:
+            self.updateLinkForces = lambda PHI, T, Norm, NormT, Bend, Twist, K, D0, Nodeinds: \
+                self.updateLinkForces3D(PHI, T, Norm, NormT, Bend, Twist, K, D0, Nodeinds)
+            self.dims = dims
+        else:
+            print "Oops! Wrong number of dimensions here."
+            sys.exit()
+
+        self.dims = dims
         # parameters for mechanical equilibration
         self.dt = dt
         self.nmax = nmax
         self.qmin = qmin
-        # 2d or 3d?
-        self.d2 = d2
-        self.is3d = 1 - self.d2
         # parameters to add/remove links
         self.d0_0 = d0_0
         self.force_limit = force_limit
@@ -107,128 +168,259 @@ class Configuration():
         self.chkx = chkx
         self.anis = anis
         self.d0max = d0max
-        # parameters to restrict mech. equilibration to surrounding of changed link
-        self.unrestricted = unrestricted
-        self.distlimit = distlimit
         # variables to store cell number and cell positions and angles
-        if num is not None:
-            self.N = num
-            self.x = np.zeros(6 * num)
-        else:
-            self.N = 0
-            self.x = None
+        self.N = num
+        self.ones = np.ones(self.N)
+        #functions for randoms in default_update_d0
+        self.lowers = np.tril_indices(self.N, -1)
+        self.randomsummand = np.zeros((self.N, self.N))
+        self.randomlength = self.N * (self.N - 1) / 2
 
-    def updateLinkForces(self, l, x):
-        l.e, l.d = norm(x[l.n2.r] - x[l.n1.r])  # actual link direction and length
-        l.n = norm(l.norm1 + l.norm2, "vec")  # Eq. 7
-        norm2 = np.cross(l.e, l.n)  # Eq. 9
-        if self.d2:
-            l.M1 = l.n1.bend * np.cross(l.e, R(x[l.n1.phi], l.t1))
-            l.M2 = l.n2.bend * np.cross(-l.e, R(x[l.n2.phi], l.t2))
-        else:
-            l.M1 = l.n1.bend * np.cross(l.e, R(x[l.n1.phi], l.t1)) + \
-                   l.n1.twist * np.cross(l.n, R(x[l.n1.phi], l.norm1))  # Eqs 3, 18, 19
-            l.M2 = l.n2.bend * np.cross(-l.e, R(x[l.n2.phi], l.t2)) + \
-                   l.n2.twist * np.cross(l.n, R(x[l.n2.phi], l.norm2))  # Eqs 5, 18, 19
-        M = l.M1 + l.M2
-        l.F2 = np.dot(M, l.n) / l.d * norm2 - np.dot(M, norm2) / l.d * l.n + l.k * (
-                l.d - l.d0) * l.e  # Eqs (11), (12) and (13) for F2=-F1!
-        l.F1 = -l.F2
+        """stuff for documentation"""
+        self.nodesnap = []
+        self.linksnap = []
+        self.fnodesnap = []
+        self.flinksnap = []
+        self.snaptimes = []
 
-    def F(self, x):
-        ret = np.zeros(len(x))  # change in state vector, 6 dof for each unit
-        for l in self.links:
-            self.updateLinkForces(l, x)
-        for n in self.nodes:
-            n.F = -n.kr * (x[n.r] - n.r0) + n.F0
-            n.M = -n.kphi * (x[n.phi] - n.phi0)
-        for l in self.links:
-            l.n1.F = l.n1.F - l.F1
-            l.n2.F = l.n2.F - l.F2
-            l.n1.M = l.n1.M - l.M1
-            l.n2.M = l.n2.M - l.M2
-        for n in self.nodes:
-            ret[n.r] = n.F
-            ret[n.phi] = n.M
-        return ret
+        """description of nodes"""
+        self.nodesX = np.zeros((self.N, 3))             # r of nodes
+        self.nodesPhi = np.zeros((self.N, 3))           # phi of nodes
+        self.Fnode = np.zeros((self.N, 3))              # total force on node
+        self.Mnode = np.zeros((self.N, 3))              # total torsion on node
+
+        """description of links"""
+        # islink[i, j] is True if nodes i and j are connected via link
+        self.islink = np.full((self.N, self.N), False)  # Describes link at node [0] leading to node [1]
+
+        self.e = np.zeros((self.N, self.N, 3))          # direction connecting nodes (a.k.a. "actual direction")
+        self.d = np.zeros((self.N, self.N))             # distance between nodes (a.k.a. "actual distance")
+        self.k = np.zeros((self.N, self.N))             # spring constant between nodes
+        self.bend = 10. * np.ones((self.N, self.N))     # bending rigidity
+        self.twist = 1. * np.ones((self.N, self.N))     # torsion spring constant
+        self.d0 = np.zeros((self.N, self.N))            # equilibrium distance between nodes,
+        self.t = np.zeros((self.N, self.N, 3))          # tangent vector of link at node (a.k.a. "preferred direction")
+        self.norm = np.zeros((self.N, self.N, 3))       # normal vector of link at node
+        self.Mlink = np.zeros((self.N, self.N, 3))      # Torsion from link on node
+        self.Flink = np.zeros((self.N, self.N, 3))      # Force from link on node
+
+        self.Qtrack = []
+        self.Ferrtrack = []
+
+    def addlink(self, ni, mi, t1=None, t2=None, d0=None, k=None, bend=None, twist=None, n=None, norm1=None, norm2=None):
+        self.islink[ni, mi], self.islink[mi, ni] = True, True
+
+        if k is None:
+            k = 15.
+        self.k[ni, mi], self.k[mi, ni] = k, k  # spring parameter
+        if bend is not None:
+            self.bend[mi, ni], self.bend[ni, mi] = bend, bend
+        if twist is not None:
+            self.twist[mi, ni], self.bend[ni, mi] = twist, twist
+
+        if d0 is None:
+            d0 = self.d[ni, mi]
+        self.d0[ni, mi], self.d0[mi, ni] = d0, d0  # equilibrium distance
+        # preferred directions
+        RotMat1 = getRotMat(-self.nodesPhi[ni])
+        RotMat2 = getRotMat(-self.nodesPhi[mi])
+        if t1 is None:
+            self.t[ni, mi] = np.dot(RotMat1, self.e[ni, mi])
+        else:
+            self.t[ni, mi] = getNormvec(t1)
+        if t2 is None:
+            self.t[mi, ni] = np.dot(RotMat2, self.e[mi, ni])
+        else:
+            self.t[mi, ni] = getNormvec(t2)
+        if n is None:
+            n, q = getNormtoo(np.cross(self.e[ni, mi], ez))  # n is perpendicular to e
+            # n is perpendicular to z (l is in the x-y plane)
+            if q < 1e-5:
+                n = getNormvec(np.cross(self.e[ni, mi], ex))  # e || ez   =>	n is perpendicular to x
+        if norm1 is None:
+            norm1 = np.dot(RotMat1, n)
+        self.norm[ni, mi] = norm1
+        if norm2 is None:
+            norm2 = np.dot(RotMat2, n)
+        self.norm[mi, ni] = norm2
+
+    def removelink(self, ni, mi):
+        self.islink[ni, mi], self.islink[mi, ni] = False, False
+        self.Flink[ni, mi], self.Flink[mi, ni], self.Mlink[ni, mi], self.Mlink[mi, ni] = null, null, null, null
+        self.t[ni, mi], self.t[mi, ni], self.norm[ni, mi], self.norm[mi, ni] = null, null, null, null
+        self.k[ni, mi], self.k[mi, ni], self.d0[ni, mi], self.d0[mi, ni] = 0, 0, 0, 0
+
+    def updateDists(self, X):
+        dX = X - X[:, None, :]
+        self.d = scipy.linalg.norm(dX, axis=2)
+        e_masked = ma.array(dX) / ma.array(self.d[..., None])
+        self.e = ma.getdata(e_masked.filled(0))
+
+    def compactStuffINeed(self):
+        # get only those parts of the big arrays that are actually needed
+        nodeinds = np.where(self.islink == True)
+        t = self.t[nodeinds]
+        norm = self.norm[nodeinds]
+        normT = np.transpose(self.norm, axes=(1, 0, 2))[nodeinds]
+        bend = self.bend[nodeinds]
+        twist = self.twist[nodeinds]
+        k = self.k[nodeinds]
+        d0 = self.d0[nodeinds]
+
+        return t, norm, normT, bend, twist, k, d0, nodeinds
+
+    def updateLinkForces2D(self, PHI, T, Norm, NormT, Bend, Twist, K, D0, Nodeinds):
+        self.Mlink.fill(0)
+        self.Flink.fill(0)
+
+        NodesPhi = PHI[Nodeinds[0]]
+        E = self.e[Nodeinds]
+        D = self.d[Nodeinds]
+        RotMat = getRotMatArray(NodesPhi)
+
+        NormNow = np.einsum("ijk, ik -> ij", RotMat, Norm)      # rotated version of norm to fit current setup
+        TNow = np.einsum("ijk, ik -> ij", RotMat, T)            # rotated version of t to fit current setup
+
+        Norm2Now = np.cross(NormNow, E)
+
+        self.Mlink[Nodeinds] = -Bend[..., None] * np.cross(TNow, E)
+
+        M = self.Mlink + np.transpose(self.Mlink, axes=(1, 0, 2))
+        M = M[Nodeinds]
+
+        inter2 = ma.array(np.einsum("ij, ij -> i", M, Norm2Now))
+        inter1 = ma.array(np.einsum("ij, ij -> i", M, NormNow))
+
+        inter1 /= ma.array(D)
+        inter2 /= ma.array(D)
+
+        inter1 = ma.getdata(ma.filled(inter1, 0))
+        inter2 = ma.getdata(ma.filled(inter2, 0))
+
+        self.Flink[Nodeinds] = -inter1[..., None] * Norm2Now + \
+                     inter2[..., None] * NormNow + \
+                     (K * (D - D0))[..., None] * E   # Eqs. 13, 14, 15
+
+    def updateLinkForces3D(self, PHI, T, Norm, NormT, Bend, Twist, K, D0, Nodeinds):
+        NodesPhi = PHI[Nodeinds[0]]
+        NodesPhiT = PHI[Nodeinds[1]]
+        E = self.e[Nodeinds]
+        D = self.d[Nodeinds]
+        RotMat = getRotMatArray(NodesPhi)
+        RotMatT = getRotMatArray(NodesPhiT)
+
+        NormNow = np.einsum("ijk, ik -> ij", RotMat, Norm)      # rotated version of norm to fit current setup
+        NormTNow = np.einsum("ijk, ik -> ij", RotMatT, NormT)
+        TNow = np.einsum("ijk, ik -> ij", RotMat, T)            # rotated version of t to fit current setup
+
+        Norm2Now = np.cross(NormNow, E)
+
+        self.Mlink[self.islink] = -Bend[..., None] * np.cross(TNow, E) + \
+                                  Twist[:, None] * np.cross(NormNow, NormTNow)  # Eqs. 5, 6
+
+        M = self.Mlink + np.transpose(self.Mlink, axes=(1, 0, 2))
+
+        inter2 = ma.array(np.einsum("ij, ij -> i", M[Nodeinds], Norm2Now))
+        inter1 = ma.array(np.einsum("ij, ij -> i", M[Nodeinds], NormNow))
+
+        inter1 /= ma.array(D)
+        inter2 /= ma.array(D)
+
+        inter1 = ma.getdata(ma.filled(inter1, 0))
+        inter2 = ma.getdata(ma.filled(inter2, 0))
+
+        self.Flink[self.islink] = -inter1[..., None] * Norm2Now + \
+                     inter2[..., None] * NormNow + \
+                     (K * (D - D0))[..., None] * E   # Eqs. 13, 14, 15
+
+    def getForces(self, X, Phi, t, norm, normT, bend, twist, k, d0, nodeinds):
+        self.updateDists(X)
+        self.updateLinkForces(Phi, t, norm, normT, bend, twist, k, d0, nodeinds)
+        self.Fnode = np.sum(self.Flink, axis=1)
+        self.Mnode = -np.sum(self.Mlink, axis=1)
+        return self.Fnode, self.Mnode
 
     def mechEquilibrium(self):
-        x = self.x.copy()
+        x = self.nodesX.copy()
+        phi = self.nodesPhi.copy()
         h = self.dt
+        steps = 0
+        t, norm, normT, bend, twist, k, d0, nodeinds = self.compactStuffINeed()
         for i in range(self.nmax):
-            k1 = self.F(x)
-            Q = np.dot(k1, k1) / len(self.nodes)
-            # print i, Q, max([x[n.r][0] for n in self.nodes])
-            if i % 100 == 0 and self.dumpdir is not None:
-                dumpData('s/%05d.pickle' % (self.dumpdir, i))
+            k1, j1 = self.getForces(x, phi, t, norm, normT, bend, twist, k, d0, nodeinds)
+            Q = (np.einsum("ij, ij", k1, k1) + np.einsum("ij, ij", j1, j1)) / self.N
             if Q < self.qmin:
                 break
-            k1 *= h
-            k2 = h * self.F(x + k1 / 2)
-            k3 = h * self.F(x + k2 / 2)
-            k4 = h * self.F(x + k3)
+            k1, j1 = h * k1, h * j1
+            k2, j2 = self.getForces(x + k1 / 2, phi + j1 / 2, t, norm, normT, bend, twist, k, d0, nodeinds)
+            k2, j2 = h * k2, h * j2
+            k3, j3 = self.getForces(x + k2 / 2, phi + j2 / 2, t, norm, normT, bend, twist, k, d0, nodeinds)
+            k3, j3 = h * k3, h * j3
+            k4, j4 = self.getForces(x + k3, phi + j3, t, norm, normT, bend, twist, k, d0, nodeinds)
+            k4, j4 = h * k4, h * j4
             x += (k1 + 2 * k2 + 2 * k3 + k4) / 6.
-        self.x = x
-        return (i + 1) * h
+            phi += (j1 + 2 * j2 + 2 * j3 + j4) / 6.
+            steps += 1
+        self.nodesX = x
+        self.nodesPhi = phi
+        return (steps + 1) * h
 
-    def delAttrDist(self):
-        for n in self.nodes:
-            if hasattr(n, 'dist'):
-                delattr(n, "dist")
+    def getLinkList(self):
+        allLinks0, allLinks1 = np.where(self.islink == True)
+        return np.array([[allLinks0[i], allLinks1[i]] for i in range(len(allLinks0)) if allLinks0[i] > allLinks1[i]])
 
     def checkLinkX(self):
         Xs = []
         delete_list = []
-        for l1, l2 in itertools.combinations(self.links, 2):
-            if intersect(l1.n1.getR(), l1.n2.getR(), l2.n1.getR(), l2.n2.getR()):
+        Links = self.getLinkList()
+        for l1, l2 in itertools.combinations(Links, 2):
+            if intersect(self.nodesX[l1[0]], self.nodesX[l1[1]], self.nodesX[l2[0]], self.nodesX[l2[1]]):
                 Xs.append([l1, l2])
-        # print "Xs:", Xs
+        print "checked for crossing, found", len(Xs)
         while len(Xs) > 0:
-            # Counter: count occurence
-            # from_iterable: unpack list of lists
-            counts = collections.Counter(itertools.chain.from_iterable(Xs))
-            # get item most in conflict (highest occurence in list)
-            l = max(counts, key=counts.get)
-            delete_list.append(l)
-            newXs = [x for x in Xs if l not in x]
+            Xsflat = np.array(Xs).reshape(2 * len(Xs), 2)
+            # u: unique elements in Xsflat (a.k.a. links), count: number of occurences in Xsflat
+            u, count = np.unique(Xsflat, axis=0, return_counts=True)
+            badlink = u[np.argmax(count)]
+            delete_list.append(badlink)
+            newXs = []
+            for linkpair in Xs:
+                if (badlink == linkpair[0]).all() or (badlink == linkpair[1]).all():
+                    continue
+                newXs.append(linkpair)
             Xs = newXs
-        for l in delete_list:
-            l.remove()
+        for badlink in delete_list:
+            self.removelink(badlink[0], badlink[1])
 
-    def delLinkList(self):
+    def delLinkList(self, linklist):
         to_del = []
-        for l in self.links:
-            if l.d < l.d0:
-                continue  # compressed links are stable
-            f = norm(l.F1, "mag")  # magnitude of force transmitted
-            if hasattr(l, "limit"):
-                p = exp(f / l.limit)
-            else:
-                p = exp(f / self.force_limit)
-            if self.anis > 1e-5:  # horizontal links are more stable
-                p = p * (1 - self.anis + self.anis * pow((l.n1.getR() - l.n2.getR())[1] / l.d, 2))  # what?
-            to_del.append((l, p))
+        for link in linklist:
+            if self.d[link[0], link[1]] < self.d0[link[0], link[1]]:
+                continue            # compressed links are stable
+            f = scipy.linalg.norm(self.Flink[link[0], link[1]])
+            p = exp(f / self.force_limit)
+            to_del.append((link, p))
         return to_del
 
-    def tryLink(self, n, m):
-        if n.findLinkTo(m) != None:
-            return -1, null
-        for l in self.links:
-            if intersect(self.x[n.r], self.x[m.r], self.x[l.n1.r], self.x[l.n2.r]):
-                return -1, null  # false
-        e, d = norm(self.x[n.r] - self.x[m.r])
+    def tryLink(self, n1, n2, Links):
+        if self.islink[n1, n2] == True:
+            return -1
+        for l in Links:
+            if intersect(self.nodesX[n1], self.nodesX[n2], self.nodesX[l[0]], self.nodesX[l[1]]):
+                print "tried link, found crossing"
+                return -1  # false
+        d = scipy.linalg.norm(self.nodesX[n1] - self.nodesX[n2])
         if d > self.d0max:
-            return -1, null  # false
-        return d, e  # true: d>0
+            return -1  # false
+        return d  # true: d>0
 
-    def addLinkList(self):
+    def addLinkList(self, Links):
         to_add = []
-        for i, j in voronoi_neighbors.VoronoiNeighbors(
-                [n.getR() for n in self.nodes], self.d0max, is3D=self.is3d):
-            d, e = self.tryLink(self.nodes[i], self.nodes[j])
+        for i, j in voronoi_neighbors.VoronoiNeighbors(self.nodesX, self.d0max, vodims=self.dims):
+            d = self.tryLink(i, j, Links)
             if d > 1e-5:
-                p = (1 - (d / 2.0)) * (1 - self.anis + self.anis * e[0] * e[0])
-                to_add.append(((self.nodes[i], self.nodes[j]), p))
+                p = (1 - (d / self.d0max))
+                to_add.append(((i, j), p))
         return to_add
 
     def pickEvent(self, to_del, to_add):
@@ -243,276 +435,66 @@ class Configuration():
         if S < 1e-7:
             print "nothing to do!"
             return 1.
-        dt = -log(random()) / S
+        dt = -log(npr.random()) / S
         if dt > 1:
             print 'Must adjust d0 variables before the next event!'
             return 1.
 
-        r = S * random()
+        r = S * npr.random()
         if r < s1:  # we will remove a link
             for (l, p) in to_del:
                 r = r - p * self.p_del
                 if r < 0:
-                    l.n1.dist = 0
-                    l.n2.dist = 0
-                    l.remove()
+                    self.removelink(l[0], l[1])
                     return dt
         r = r - s1
         if r < s2:  # we will add a link
             for ((n1, n2), p) in to_add:
                 r = r - p * self.p_add
                 if r < 0:
-                    n1.dist = 0
-                    n2.dist = 0
-                    n1.addLinkTo(n2)
+                    self.addlink(n1, n2)
                     return dt
 
     def default_update_d0(self, dt):
-        for l in self.links:
-            l.d0 += 0.2 * (self.d0_0 - l.d0) * dt + 0.05 * (
-                        2 * sqrt(dt) * random() - sqrt(dt))  # magic number 0.2 and 0.05??
+        myrandom = npr.random((self.randomlength, ))
+        self.randomsummand[self.lowers], self.randomsummand.T[self.lowers] = myrandom, myrandom
+        self.d0 += 0.2 * (self.d0_0 - self.d0) * dt + 0.05 * (
+                   2 * sqrt(dt) * self.randomsummand - sqrt(dt))              # magic number 0.2 and 0.05??
 
     def modlink(self):
-        self.delAttrDist()
         if self.chkx:
             self.checkLinkX()
-
-        to_del = self.delLinkList()
-        to_add = self.addLinkList()
+        Links = self.getLinkList()
+        to_del = self.delLinkList(Links)
+        to_add = self.addLinkList(Links)
         dt = self.pickEvent(to_del, to_add)
         self.default_update_d0(dt)
         return dt
 
+    def makesnap(self, t):
+        self.nodesnap.append(self.nodesX.copy())
+        self.fnodesnap.append(self.Fnode.copy())
+        linkList = self.getLinkList()
+        self.linksnap.append(linkList)
+        self.flinksnap.append(scipy.linalg.norm(self.Flink[linkList[..., 0], linkList[..., 1]], axis = 1))
+        self.snaptimes.append(t)
+
     def timeevo(self, tmax, record=False):
-        # configs, ts = [copy.deepcopy(self)], [0.]
         t = 0.
+        if record:
+            self.makesnap(0)
         while t < tmax:
             dt = self.mechEquilibrium()
             t += dt
             if record:
-                configs.append(copy.deepcopy(self))
-                ts.append(t)
-            dt = self.modlink()
+                self.makesnap(t)
             update_progress(t / tmax)
+            dt = self.modlink()
             t += dt
             if record:
-                configs.append(copy.deepcopy(self))
-                ts.append(t)
+                self.makesnap(t)
             update_progress(t / tmax)
         if record:
-            return configs, ts
-
-
-def showconfig(config, figure=None, figureindex=0, bgcolor=(1, 1, 1), fgcolor=(0, 0, 0), figsize=(1000, 1000),
-               cmap='viridis', vmaxlinks=5, vmaxcells=5,
-               cbar=False):
-    if figure is None:
-        fig = mlab.figure(figureindex, bgcolor=bgcolor, fgcolor=fgcolor, size=figsize)
-    else:
-        fig = figure
-    x, y, z, phi1, phi2, phi3 = config.x.reshape(6, config.N, order='F')
-    posl, rl, dl, fl, fc = [], [], [], [], []
-    for l in config.links:
-        posl.append(config.x[l.n1.r])
-        rl.append(l.e)
-        dl.append(l.d)
-        fl.append(norm(l.F1, 'mag'))
-
-    for c in config.nodes:
-        fc.append(norm(c.F, 'mag'))
-
-    dl = np.asarray(dl)
-    fl = np.asarray(fl)
-    xl, yl, zl = np.asarray(posl).T
-    rxl, ryl, rzl = np.asarray(rl).T
-    rxl *= dl
-    ryl *= dl
-    rzl *= dl
-
-    cells = mlab.points3d(x, y, z, fc, scale_factor=1, opacity=0.5, resolution=16, scale_mode='none', vmin=0.,
-                          colormap=cmap, vmax=vmaxcells)
-    links = mlab.quiver3d(xl, yl, zl, rxl, ryl, rzl, scalars=fl, mode='2ddash', line_width=4., scale_mode='vector',
-                          scale_factor=1, colormap=cmap, vmin=0., vmax=vmaxlinks)
-    links.glyph.color_mode = "color_by_scalar"
-    if cbar:
-        mlab.scalarbar(links, nb_labels=2, title='Force on link')
-    # mlab.draw()
-    return cells, links
-
-
-@mlab.animate(delay=500)
-def animateconfigs(configs, ts=None, figureindex=0, bgcolor=(1, 1, 1), fgcolor=(0, 0, 0), figsize=(1000, 1000),
-                   cmap='viridis', cbar=False):
-    fig = mlab.figure(figureindex, bgcolor=bgcolor, fgcolor=fgcolor, size=figsize)
-    c = configs[0]
-    vmaxcells = max([max([norm(node.F, 'mag') for node in config.nodes]) for config in configs])
-    vmaxlinks = max([max([norm(link.F1, 'mag') for link in config.links]) for config in configs])
-    cells, links = showconfig(c, figure=fig, vmaxcells=vmaxcells, vmaxlinks=vmaxlinks)
-    text = mlab.title('0.0', height=.9)  # (cells.mlab_source.x.max(), cells.mlab_source.y.max(), cells.mlab_source.z.max(), '0.0')
-    if ts is None:
-        ts = range(len(configs))
-    while True:
-        for (c, t) in zip(configs, ts):
-            x, y, z, phi1, phi2, phi3 = c.x.reshape(6, c.N, order='F')
-            posl, rl, dl, fl, fc = [], [], [], [], []
-            for l in c.links:
-                posl.append(c.x[l.n1.r])
-                rl.append(l.e)
-                dl.append(l.d)
-                fl.append(norm(l.F1, 'mag'))
-
-            for n in c.nodes:
-                fc.append(norm(n.F, 'mag'))
-
-            dl = np.asarray(dl)
-            fl = np.asarray(fl)
-            xl, yl, zl = np.asarray(posl).T
-            rxl, ryl, rzl = np.asarray(rl).T
-            rxl *= dl
-            ryl *= dl
-            rzl *= dl
-            cells.mlab_source.set(x=x, y=y, z=z, scalars=fc)
-            links.mlab_source.reset(x=xl, y=yl, z=zl, u=rxl, v=ryl, w=rzl, scalars=fl)
-            text.set(text='{}'.format(round(t, 2)))
-            print 'Updating... '
-            yield
-
-class node(object):
-    def __init__(self, config, r, r0=None, state=-1):
-
-        self.config = config  # reference to a specific configuration
-        # state vector indices
-        if config.nNodes == len(config.x):
-            config.N += 1
-            config.x.append(np.zeros(6))
-
-        self.r = [config.nNodes, config.nNodes + 1, config.nNodes + 2]
-        config.nNodes += 3
-        config.x[self.r] = r
-        if r0 is None:
-            r0 = r
-        self.r0 = r0
-        # state vector indices
-        self.phi = [config.nNodes, config.nNodes + 1, config.nNodes + 2]
-        config.nNodes += 3
-        config.x[self.phi] = null
-        self.phi0 = null
-
-        self.links = []
-        self.F = null  # net force
-        self.F0 = null  # external force
-        self.M = null  # net torque
-
-        self.bend = 10.  # bending rigidity
-        self.twist = 1.  # torsion spring constant
-        self.kr = 0.  # spring to anchor point
-        self.kphi = 0.  # torsion spring to anchor point
-
-        self.state = state  # discrete state variable, like 0, 1, 2, ..
-        self.normal = None  # normal vector, if surface node
-        config.nodes.append(self)
-
-    def findLinkTo(self, n):
-        for l in self.links:
-            if ((l.n1 == self) & (l.n2 == n)) | ((l.n1 == n) & (l.n2 == self)):
-                return l
-        return None
-
-    def addLinkTo(self, n, t1=None, t2=None, d0=None, k=None):
-        if self.findLinkTo(n) is None:
-            l = link(self.config, self, n, t1, t2, d0, k)
-            self.links.append(l)
-            n.links.append(l)
-            return l
-
-    def rmLinkTo(self, n):
-        l = self.findLinkTo(n)
-        if l is not None:
-            self.links.remove(l)
-            n.links.remove(l)
-            self.config.links.remove(l)
-
-    def getR(self):
-        return self.config.x[self.r]
-
-
-class link(object):
-    def __init__(self, config, n1, n2, t1=None, t2=None, d0=None, k=None, n=None, norm1=None, norm2=None):
-
-        self.config = config  # reference to a specific configuration
-        self.n1, self.n2 = n1, n2  # nodes at endpoints
-        # actual direction e and length d
-        self.e, self.d = norm(config.x[n2.r] - config.x[n1.r])  # vector and its length
-
-        if k is None:
-            k = 15.
-        self.k = k  # spring parameter
-
-        if d0 is None:
-            d0 = self.d
-        self.d0 = d0  # equilibrium distance
-        # preferred directions
-        if t1 is None:
-            self.t1 = R(-config.x[self.n1.phi], self.e)
-        else:
-            self.t1 = norm(t1, "vec")
-        if t2 is None:
-            self.t2 = R(-config.x[self.n2.phi], -self.e)
-        else:
-            self.t2 = norm(t2, "vec")
-        if n is None:
-            n, q = norm(np.cross(self.e, ez))  # n is perpendicular to e
-            # n is perpendicular to z (l is in the x-y plane)
-            if q < 1e-5:
-                n = norm(np.cross(self.e, ex), "vec")  # e || ez   =>	n is perpendicular to x
-        self.n = n  # perpendicular vector to link
-        if norm1 is None:
-            norm1 = n
-        self.norm1 = norm1
-        if norm2 is None:
-            norm2 = n
-        self.norm2 = norm2
-        self.M1, self.M2, self.F1, self.F2 = null, null, null, null
-        config.links.append(self)
-
-    def remove(self):
-        self.n1.links.remove(self)
-        self.n2.links.remove(self)
-        self.config.links.remove(self)
-
-
-def checkDir(s):
-    if os.path.exists(s):
-        if not os.path.isdir(s):
-            raise argparse.ArgumentTypeError("%s is not a dir." % s)
-    else:
-        os.makedirs(s)
-    return s
-
-
-def checkFile(s):
-    if os.path.exists(s):
-        if s.strip() == '/dev/stdin':
-            return s
-        if not os.path.isfile(s):
-            raise argparse.ArgumentTypeError("%s is not a file." % s)
-        return s
-    else:
-        argparse.ArgumentTypeError("%s does not exists." % s)
-
-
-def dumpData(configuration, fname):
-    sys.setrecursionlimit(50000)
-    with gzip.GzipFile(fname, 'wb') as out:
-        pickle.dump(configuration, out)
-
-
-def loadData(fname):
-    with gzip.GzipFile(fname, 'rb') as inf:
-        config = pickle.load(inf)
-
-    print >> sys.stderr, "N=", config.N
-    print >> sys.stderr, len(config.x), "dof."
-    print >> sys.stderr, len(config.nodes), "nodes."
-    print >> sys.stderr, len(config.links), "links."
-    return config
+            self.nodesnap = np.array(self.nodesnap)
+            self.fnodesnap = np.array(self.fnodesnap)
+            return self.nodesnap, self.linksnap, self.fnodesnap, self.flinksnap, self.snaptimes
