+++ minimal.py
@@ -4,16 +4,15 @@ from cell import *
 import cProfile
 import matplotlib.pyplot as plt
 
+npr.seed(seed=0)
 
 #######################################################
 
 def generate_config_from_default(R):
     N = len(R)
-    c = Configuration(num=N, d2=True)
+    c = Configuration(N, dims=2)
     for ni in range(N):
-        n = node(c, R[ni])
-        n.twist = twist
-        n.bend = bend
+        c.nodesX[ni] = R[ni]
 
     return c
 
@@ -35,36 +34,59 @@ if __name__ == '__main__':
     anis = 1.0  # anisotropy of building links (we don't need this so far)
     chkx = True  # check if links overlap?
 
-    N = 4
+    N = 3
 
     # config = generate_initial_config(Lmax)
 
     xpos = np.sqrt(2 * 1.3 * 1.3) / 2.
 
-    R = np.array([[-xpos, xpos, 0], [0, 0., 0], [xpos, xpos, 0], [0, 2 * xpos, 0]])
+    # R = np.array([[-xpos, xpos, 0], [0, 0., 0], [xpos, xpos, 0], [0, 2 * xpos, 0]])
+    R = np.array([[0, 0, 0], [1, 0, 0], [1, 1, 0], [0, 1, 0]])
     config = generate_config_from_default(R)
 
-    for i in range(N):
-        config.nodes[i].addLinkTo(config.nodes[(i + 1)%N], d0 = 1.)
+    config.updateDists(config.nodesX)
 
-    configs, ts = [copy.deepcopy(config)], [0.]
+    config.addlink(0, 2)
+    config.addlink(1, 3)
+    # config.addlink(2, 3)
+    # config.addlink(3, 0)
+
+    # config.addlink(0, 2)
+
+    config.checkLinkX()
+    # configs, links, nodeforces, linkforces, ts = config.minitimeevo(4., record=True, now=False)
+    """
     t = 0.
-    dt = config.mechEquilibrium()
-    t += dt
-    configs.append(copy.deepcopy(config))
-    ts.append(t)
+    config.makesnap(0)
 
-    config.nodes[0].addLinkTo(config.nodes[2], d0 = 2.)
+    thisdt = config.mechEquilibrium(now=False)
+    print "equil: dt = ", thisdt
+    t += thisdt
+    print "t = ", t
+    config.makesnap(t)
 
-    dt = config.mechEquilibrium()
-    t += dt
-    configs.append(copy.deepcopy(config))
-    ts.append(t)
 
-    # Qtie = np.linspace(1, len(config.Qtrack), len(config.Qtrack))
-    # plt.plot(Qtie, config.Qtrack)
+    myd0 = scipy.linalg.norm(config.nodesX[0] - config.nodesX[2])
+
+
+
+    config.addlink(0, 2, d0 = 2.)
+    config.updateDists(config.nodesX)
+
+    thisdt = config.mechEquilibrium(now=False)
+    print "equil: dt = ", thisdt
+    t += thisdt
+    print "t = ", t
+    config.makesnap(t)
+
+    Qtie = np.linspace(1, len(config.Qtrack), len(config.Qtrack))
+    plt.plot(Qtie, config.Qtrack, color="blue")
+    Ferrtie = np.linspace(1, len(config.Ferrtrack), len(config.Ferrtrack))
+    plt.plot(Ferrtie, config.Ferrtrack, color="red")
     # plt.show()
 
-
-    animateconfigs(configs, ts=ts)
+    config.nodesnap = np.array(config.nodesnap)
+    config.fnodesnap = np.array(config.fnodesnap)
+    """
+    animateconfigs(config.nodesnap, config.linksnap, config.fnodesnap, config.flinksnap, config.snaptimes)
     mlab.show()
